separating the advanced search dialog, remove unused old dialog
