+++ megamek/src/megamek/client/ui/advancedsearch/EquipmentTableModel.java
@@ -1,0 +1,140 @@
+package megamek.client.ui.advancedsearch;
+
+import megamek.common.EquipmentType;
+import megamek.common.TechConstants;
+
+import javax.swing.table.AbstractTableModel;
+import java.util.Arrays;
+import java.util.Vector;
+
+/**
+ * A table model for displaying equipment
+ */
+public class EquipmentTableModel extends AbstractTableModel {
+
+    static final int COL_QTY = 0;
+    static final int COL_NAME = 1;
+    static final int COL_COST = 2;
+    static final int COL_IS_CLAN = 3;
+    static final int COL_LEVEL = 4;
+    static final int N_COL = 5;
+    static final int COL_INTERNAL_NAME = 5;
+
+    private final TWAdvancedSearchPanel twAdvancedSearchPanel;
+    private int[] qty;
+    private Vector<EquipmentType> equipment = new Vector<>();
+
+    public EquipmentTableModel(TWAdvancedSearchPanel twAdvancedSearchPanel) {
+        this.twAdvancedSearchPanel = twAdvancedSearchPanel;
+    }
+
+    @Override
+    public int getRowCount() {
+        return equipment.size();
+    }
+
+    @Override
+    public int getColumnCount() {
+        return N_COL;
+    }
+
+    public int getPreferredWidth(int col) {
+        switch (col) {
+            case COL_QTY:
+                return 40;
+            case COL_NAME:
+                return 400;
+            case COL_IS_CLAN:
+                return 75;
+            case COL_COST:
+                return 175;
+            case COL_LEVEL:
+                return 100;
+            default:
+                return 0;
+        }
+    }
+
+    @Override
+    public String getColumnName(int column) {
+        switch (column) {
+            case COL_QTY:
+                return "Qty";
+            case COL_NAME:
+                return "Name";
+            case COL_IS_CLAN:
+                return "IS/Clan";
+            case COL_COST:
+                return "Cost";
+            case COL_LEVEL:
+                return "Lvl";
+            default:
+                return "?";
+        }
+    }
+
+    @Override
+    public Class<?> getColumnClass(int c) {
+        return getValueAt(0, c).getClass();
+    }
+
+    @Override
+    public boolean isCellEditable(int row, int col) {
+        switch (col) {
+            case COL_QTY:
+                return true;
+            default:
+                return false;
+        }
+    }
+
+    // fill table with values
+    public void setData(Vector<EquipmentType> eq) {
+        equipment = eq;
+        qty = new int[eq.size()];
+        Arrays.fill(qty, 1);
+        fireTableDataChanged();
+    }
+
+    public EquipmentType getEquipmentTypeAt(int row) {
+        return equipment.elementAt(row);
+    }
+
+    @Override
+    public Object getValueAt(int row, int col) {
+        if (row >= equipment.size()) {
+            return null;
+        }
+        EquipmentType eq = equipment.elementAt(row);
+        switch (col) {
+            case COL_QTY:
+                return qty[row] + "";
+            case COL_NAME:
+                return eq.getName();
+            case COL_IS_CLAN:
+                return TechConstants.getTechName(eq.getTechLevel(twAdvancedSearchPanel.gameYear));
+            case COL_COST:
+                return eq.getRawCost();
+            case COL_LEVEL:
+                return TechConstants.getSimpleLevelName(TechConstants
+                    .convertFromNormalToSimple(eq
+                        .getTechLevel(twAdvancedSearchPanel.gameYear)));
+            case COL_INTERNAL_NAME:
+                return eq.getInternalName();
+            default:
+                return "?";
+        }
+    }
+
+    @Override
+    public void setValueAt(Object value, int row, int col) {
+        switch (col) {
+            case COL_QTY:
+                qty[row] = Integer.parseInt((String) value);
+                fireTableCellUpdated(row, col);
+                break;
+            default:
+                break;
+        }
+    }
+}
