+++ megamek/src/megamek/client/ui/advancedsearch/FilterTokens.java
@@ -1,0 +1,10 @@
+package megamek.client.ui.advancedsearch;
+
+/**
+ * Base class for different tokens that can be in a filter expression.
+ *
+ * @author Arlith
+ */
+public class FilterTokens {
+
+}
