+++ megamek/src/megamek/client/ui/advancedsearch/NoSelectionModel.java
@@ -1,0 +1,21 @@
+package megamek.client.ui.advancedsearch;
+
+import javax.swing.*;
+
+class NoSelectionModel extends DefaultListSelectionModel {
+    @Override
+    public void setAnchorSelectionIndex(final int anchorIndex) {
+    }
+
+    @Override
+    public void setLeadAnchorNotificationEnabled(final boolean flag) {
+    }
+
+    @Override
+    public void setLeadSelectionIndex(final int leadIndex) {
+    }
+
+    @Override
+    public void setSelectionInterval(final int index0, final int index1) {
+    }
+}
