+++ megamek/src/megamek/client/ui/advancedsearch/ParensFT.java
@@ -1,0 +1,19 @@
+package megamek.client.ui.advancedsearch;
+
+/**
+ * FilterTokens subclass that represents parenthesis.
+ *
+ * @author Arlith
+ */
+public class ParensFT extends FilterTokens {
+    public String parens;
+
+    public ParensFT(String p) {
+        parens = p;
+    }
+
+    @Override
+    public String toString() {
+        return parens;
+    }
+}
