+++ megamek/src/megamek/client/ui/advancedsearch/TWAdvancedSearchPanel.java
@@ -17,10 +17,9 @@
  * You should have received a copy of the GNU General Public License
  * along with MegaMek. If not, see <http://www.gnu.org/licenses/>.
  */
-package megamek.client.ui.swing.unitSelector;
+package megamek.client.ui.advancedsearch;
 
 import java.awt.BorderLayout;
-import java.awt.Dimension;
 import java.awt.FlowLayout;
 import java.awt.Font;
 import java.awt.GridBagConstraints;
@@ -34,7 +33,6 @@ import java.awt.event.KeyListener;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Enumeration;
 import java.util.List;
 import java.util.Map;
@@ -45,14 +43,12 @@ import javax.swing.*;
 import javax.swing.border.Border;
 import javax.swing.event.ListSelectionEvent;
 import javax.swing.event.ListSelectionListener;
-import javax.swing.table.AbstractTableModel;
 import javax.swing.table.TableColumn;
 import javax.swing.table.TableRowSorter;
 
 import megamek.MMConstants;
 import megamek.client.ui.Messages;
 import megamek.client.ui.swing.table.MegaMekTable;
-import megamek.client.ui.swing.util.UIUtil;
 import megamek.common.*;
 import megamek.common.equipment.ArmorType;
 import megamek.common.options.AbstractOptions;
@@ -379,7 +375,7 @@ public class TWAdvancedSearchPanel extends JPanel implements ActionListener, Ite
     private JButton btnFilterDoomedInVacuum = new JButton("\u2610");
 
     /** The game's current year. */
-    private int gameYear;
+    int gameYear;
 
     /**
      * Constructs a new advanced search panel for Total Warfare values
@@ -413,42 +409,6 @@ public class TWAdvancedSearchPanel extends JPanel implements ActionListener, Ite
         twSearchPane.addTab(msg_quirkType, quirkPanel);
 
         this.add(twSearchPane, BorderLayout.NORTH);
-    }
-
-    private static class NoSelectionModel extends DefaultListSelectionModel {
-        @Override
-        public void setAnchorSelectionIndex(final int anchorIndex) {}
-
-        @Override
-        public void setLeadAnchorNotificationEnabled(final boolean flag) {}
-
-        @Override
-        public void setLeadSelectionIndex(final int leadIndex) {}
-
-        @Override
-        public void setSelectionInterval(final int index0, final int index1) {}
-    }
-
-    private static class TriStateItem {
-        public String state;
-        public String text;
-        public int code;
-
-        public TriStateItem(String state, String text) {
-            this.state = state;
-            this.text = text;
-        }
-
-        public TriStateItem(String state, int code, String text) {
-            this.state = state;
-            this.code = code;
-            this.text = text;
-        }
-
-        @Override
-        public String toString() {
-            return state + " " + text;
-        }
     }
 
     private void jListSetup(JList<TriStateItem> l) {
@@ -1369,7 +1329,7 @@ public class TWAdvancedSearchPanel extends JPanel implements ActionListener, Ite
         scrTableWeaponType.setViewportView(tblWeaponType);
 
         // Setup Weapons Table
-        weaponsModel = new WeaponsTableModel();
+        weaponsModel = new WeaponsTableModel(this);
         tblWeapons = new MegaMekTable(weaponsModel, WeaponsTableModel.COL_NAME);
         TableColumn wpsCol = tblWeapons.getColumnModel().getColumn(WeaponsTableModel.COL_QTY);
         wpsCol.setCellEditor(new DefaultCellEditor(cboQty));
@@ -1387,7 +1347,7 @@ public class TWAdvancedSearchPanel extends JPanel implements ActionListener, Ite
         scrTableWeapons.setViewportView(tblWeapons);
 
         // Setup Equipment Table
-        equipmentModel = new EquipmentTableModel();
+        equipmentModel = new EquipmentTableModel(this);
         tblEquipment = new MegaMekTable(equipmentModel, EquipmentTableModel.COL_NAME);
         TableColumn eqCol = tblEquipment.getColumnModel().getColumn(EquipmentTableModel.COL_QTY);
         eqCol.setCellEditor(new DefaultCellEditor(cboQty));
@@ -2448,404 +2408,6 @@ public class TWAdvancedSearchPanel extends JPanel implements ActionListener, Ite
         }
     }
 
-    public class WeaponClassTableModel extends AbstractTableModel {
-        private static final long serialVersionUID = 1L;
-
-        private static final int COL_QTY = 0;
-        private static final int COL_NAME = 1;
-        private static final int N_COL = 2;
-        private static final int COL_VAL = 2;
-
-
-        private int[] qty;
-
-        private Vector<WeaponClass> weaponClasses = new Vector<>();
-
-        public WeaponClassTableModel() {
-            for (WeaponClass cl : WeaponClass.values())
-            {
-                weaponClasses.add(cl);
-            }
-            qty = new int[weaponClasses.size()];
-            Arrays.fill(qty, 1);
-        }
-
-        @Override
-        public int getRowCount() {
-            return weaponClasses.size();
-        }
-
-        @Override
-        public int getColumnCount() {
-            return N_COL;
-        }
-
-        public int getPreferredWidth(int col) {
-            switch (col) {
-                case COL_QTY:
-                    return 40;
-                case COL_NAME:
-                    return 310;
-                default:
-                    return 0;
-            }
-        }
-
-        @Override
-        public String getColumnName(int column) {
-            switch (column) {
-                case COL_QTY:
-                    return "Qty";
-                case COL_NAME:
-                    return "Weapon Class";
-                default:
-                    return "?";
-            }
-        }
-
-        @Override
-        public Class<?> getColumnClass(int c) {
-            return getValueAt(0, c).getClass();
-        }
-
-        @Override
-        public boolean isCellEditable(int row, int col) {
-            switch (col) {
-                case COL_QTY:
-                    return true;
-                default:
-                    return false;
-            }
-        }
-
-        // fill table with values
-        // public void setData(Vector<Integer> wps) {
-        //     weaponClasses = wps;
-        //     qty = new int[wps.size()];
-        //     Arrays.fill(qty, 1);
-        //     fireTableDataChanged();
-        // }
-
-        public WeaponClass getWeaponTypeAt(int row) {
-            return weaponClasses.elementAt(row);
-        }
-
-        @Override
-        public Object getValueAt(int row, int col) {
-            if (row >= weaponClasses.size()) {
-                return null;
-            }
-
-            switch (col) {
-                case COL_QTY:
-                    return qty[row] + "";
-                case COL_NAME:
-                    return weaponClasses.elementAt(row).toString();
-                case COL_VAL:
-                    return weaponClasses.elementAt(row);
-                default:
-                    return "?";
-            }
-        }
-
-        @Override
-        public void setValueAt(Object value, int row, int col) {
-            switch (col) {
-                case COL_QTY:
-                    qty[row] = Integer.parseInt((String) value);
-                    fireTableCellUpdated(row, col);
-                    break;
-                default:
-                    break;
-            }
-        }
-
-    }
-
-
-    /**
-     * A table model for displaying weapons
-     */
-    public class WeaponsTableModel extends AbstractTableModel {
-
-        private static final int COL_QTY = 0;
-        private static final int COL_NAME = 1;
-        private static final int COL_DMG = 2;
-        private static final int COL_HEAT = 3;
-        private static final int COL_SHORT = 4;
-        private static final int COL_MED = 5;
-        private static final int COL_LONG = 6;
-        private static final int COL_IS_CLAN = 7;
-        private static final int COL_LEVEL = 8;
-        private static final int N_COL = 9;
-        private static final int COL_INTERNAL_NAME = 9;
-
-        private int[] qty;
-
-        private Vector<WeaponType> weapons = new Vector<>();
-
-        @Override
-        public int getRowCount() {
-            return weapons.size();
-        }
-
-        @Override
-        public int getColumnCount() {
-            return N_COL;
-        }
-
-        public int getPreferredWidth(int col) {
-            switch (col) {
-                case COL_QTY:
-                    return 40;
-                case COL_NAME:
-                    return 310;
-                case COL_IS_CLAN:
-                    return 75;
-                case COL_DMG:
-                    return 50;
-                case COL_HEAT:
-                    return 50;
-                case COL_SHORT:
-                    return 50;
-                case COL_MED:
-                    return 50;
-                case COL_LONG:
-                    return 50;
-                case COL_LEVEL:
-                    return 100;
-                default:
-                    return 0;
-            }
-        }
-
-        @Override
-        public String getColumnName(int column) {
-            switch (column) {
-                case COL_QTY:
-                    return "Qty";
-                case COL_NAME:
-                    return "Weapon Name";
-                case COL_IS_CLAN:
-                    return "IS/Clan";
-                case COL_DMG:
-                    return "DMG";
-                case COL_HEAT:
-                    return "Heat";
-                case COL_SHORT:
-                    return "Short";
-                case COL_MED:
-                    return "Med";
-                case COL_LONG:
-                    return "Long";
-                case COL_LEVEL:
-                    return "Lvl";
-                default:
-                    return "?";
-            }
-        }
-
-        @Override
-        public Class<?> getColumnClass(int c) {
-            return getValueAt(0, c).getClass();
-        }
-
-        @Override
-        public boolean isCellEditable(int row, int col) {
-            switch (col) {
-                case COL_QTY:
-                    return true;
-                default:
-                    return false;
-            }
-        }
-
-        // fill table with values
-        public void setData(Vector<WeaponType> wps) {
-            weapons = wps;
-            qty = new int[wps.size()];
-            Arrays.fill(qty, 1);
-            fireTableDataChanged();
-        }
-
-        public WeaponType getWeaponTypeAt(int row) {
-            return weapons.elementAt(row);
-        }
-
-        @Override
-        public Object getValueAt(int row, int col) {
-            if (row >= weapons.size()) {
-                return null;
-            }
-            WeaponType wp = weapons.elementAt(row);
-            switch (col) {
-                case COL_QTY:
-                    return qty[row] + "";
-                case COL_NAME:
-                    return wp.getName();
-                case COL_IS_CLAN:
-                    return TechConstants.getTechName(wp.getTechLevel(gameYear));
-                case COL_DMG:
-                    return wp.getDamage();
-                case COL_HEAT:
-                    return wp.getHeat();
-                case COL_SHORT:
-                    return wp.getShortRange();
-                case COL_MED:
-                    return wp.getMediumRange();
-                case COL_LONG:
-                    return wp.getLongRange();
-                case COL_LEVEL:
-                    return TechConstants.getSimpleLevelName(TechConstants
-                            .convertFromNormalToSimple(wp
-                                    .getTechLevel(gameYear)));
-                case COL_INTERNAL_NAME:
-                    return wp.getInternalName();
-                default:
-                    return "?";
-            }
-        }
-
-        @Override
-        public void setValueAt(Object value, int row, int col) {
-            switch (col) {
-                case COL_QTY:
-                    qty[row] = Integer.parseInt((String) value);
-                    fireTableCellUpdated(row, col);
-                    break;
-                default:
-                    break;
-            }
-        }
-
-    }
-
-    /**
-     * A table model for displaying equipment
-     */
-    public class EquipmentTableModel extends AbstractTableModel {
-
-        private static final int COL_QTY = 0;
-        private static final int COL_NAME = 1;
-        private static final int COL_COST = 2;
-        private static final int COL_IS_CLAN = 3;
-        private static final int COL_LEVEL = 4;
-        private static final int N_COL = 5;
-        private static final int COL_INTERNAL_NAME = 5;
-
-        private int[] qty;
-        private Vector<EquipmentType> equipment = new Vector<>();
-
-        @Override
-        public int getRowCount() {
-            return equipment.size();
-        }
-
-        @Override
-        public int getColumnCount() {
-            return N_COL;
-        }
-
-        public int getPreferredWidth(int col) {
-            switch (col) {
-                case COL_QTY:
-                    return 40;
-                case COL_NAME:
-                    return 400;
-                case COL_IS_CLAN:
-                    return 75;
-                case COL_COST:
-                    return 175;
-                case COL_LEVEL:
-                    return 100;
-                default:
-                    return 0;
-            }
-        }
-
-        @Override
-        public String getColumnName(int column) {
-            switch (column) {
-                case COL_QTY:
-                    return "Qty";
-                case COL_NAME:
-                    return "Name";
-                case COL_IS_CLAN:
-                    return "IS/Clan";
-                case COL_COST:
-                    return "Cost";
-                case COL_LEVEL:
-                    return "Lvl";
-                default:
-                    return "?";
-            }
-        }
-
-        @Override
-        public Class<?> getColumnClass(int c) {
-            return getValueAt(0, c).getClass();
-        }
-
-        @Override
-        public boolean isCellEditable(int row, int col) {
-            switch (col) {
-                case COL_QTY:
-                    return true;
-                default:
-                    return false;
-            }
-        }
-
-        // fill table with values
-        public void setData(Vector<EquipmentType> eq) {
-            equipment = eq;
-            qty = new int[eq.size()];
-            Arrays.fill(qty, 1);
-            fireTableDataChanged();
-        }
-
-        public EquipmentType getEquipmentTypeAt(int row) {
-            return equipment.elementAt(row);
-        }
-
-        @Override
-        public Object getValueAt(int row, int col) {
-            if (row >= equipment.size()) {
-                return null;
-            }
-            EquipmentType eq = equipment.elementAt(row);
-            switch (col) {
-                case COL_QTY:
-                    return qty[row] + "";
-                case COL_NAME:
-                    return eq.getName();
-                case COL_IS_CLAN:
-                    return TechConstants.getTechName(eq.getTechLevel(gameYear));
-                case COL_COST:
-                    return eq.getRawCost();
-                case COL_LEVEL:
-                    return TechConstants.getSimpleLevelName(TechConstants
-                            .convertFromNormalToSimple(eq
-                                    .getTechLevel(gameYear)));
-                case COL_INTERNAL_NAME:
-                    return eq.getInternalName();
-                default:
-                    return "?";
-            }
-        }
-
-        @Override
-        public void setValueAt(Object value, int row, int col) {
-            switch (col) {
-                case COL_QTY:
-                    qty[row] = Integer.parseInt((String) value);
-                    fireTableCellUpdated(row, col);
-                    break;
-                default:
-                    break;
-            }
-        }
-    }
 
     @Override
     public void keyPressed(KeyEvent evt) {
@@ -2874,264 +2436,4 @@ public class TWAdvancedSearchPanel extends JPanel implements ActionListener, Ite
     }
 
 
-    /**
-     * Base class for different tokens that can be in a filter expression.
-     * @author Arlith
-     */
-    public static class FilterTokens {
-
-    }
-
-    /**
-     * FilterTokens subclass that represents parenthesis.
-     * @author Arlith
-     */
-    public static class ParensFT extends FilterTokens {
-        public String parens;
-
-        public ParensFT(String p) {
-            parens = p;
-        }
-
-        @Override
-        public String toString() {
-            return parens;
-        }
-    }
-
-    /**
-     * FilterTokens subclass that represents equipment.
-     * @author Arlith
-     */
-    public static class EquipmentFT extends FilterTokens {
-        public String internalName;
-        public String fullName;
-        public int qty;
-
-        public EquipmentFT(String in, String fn, int q) {
-            internalName = in;
-            fullName = fn;
-            qty = q;
-        }
-
-        @Override
-        public String toString() {
-            return qty + " " + fullName + ((qty != 1) ? "s" : "");
-        }
-    }
-
-    /**
-     * FilterTokens subclass that represents a boolean operation.
-     * @author Arlith
-     */
-    public static class OperationFT extends FilterTokens {
-        public MekSearchFilter.BoolOp op;
-
-        public OperationFT(MekSearchFilter.BoolOp o) {
-            op = o;
-        }
-
-        @Override
-        public String toString() {
-            if (op == MekSearchFilter.BoolOp.AND) {
-                return "And";
-            } else if (op == MekSearchFilter.BoolOp.OR) {
-                return "Or";
-            } else {
-                return "";
-            }
-        }
-    }
-
-    public class WeaponClassFT extends FilterTokens {
-        public WeaponClass weaponClass;
-        public int qty;
-
-        public WeaponClassFT(WeaponClass in_class, int in_qty) {
-            weaponClass = in_class;
-            qty = in_qty;
-        }
-
-        @Override
-        public String toString() {
-            if (qty == 1) {
-                return qty + " " + weaponClass.toString();
-            } else {
-                return qty + " " + weaponClass.toString() + "s";
-            }
-        }
-    }
-
-    public static enum WeaponClass {
-        AUTOCANNON {
-            public String toString() {
-                return "Autocannon";
-            }
-        },
-        RAC,
-        ULTRA {
-            public String toString() {
-                return "Ultra A/C";
-            }
-        },
-        LIGHT {
-            public String toString() {
-                return "Light A/C";
-            }
-        },
-        MACHINE_GUN {
-            public String toString() {
-                return "Machine Gun";
-            }
-        },
-        GAUSS {
-            public String toString() {
-                return "Gauss";
-            }
-        },
-        BALLISTIC {
-            public String toString() {
-                return "Ballistic";
-            }
-        },
-        PLASMA {
-            public String toString() {
-                return "Plasma";
-            }
-        },
-        ENERGY {
-            public String toString() {
-                return "Energy";
-            }
-        },
-        LASER {
-            public String toString() {
-                return "Laser";
-            }
-        },
-        PULSE {
-            public String toString() {
-                return "Pulse Laser";
-            }
-        },
-        RE_ENGINEERED {
-            public String toString() {
-                return "Re-Engineered Laser";
-            }
-        },
-        PPC {
-            public String toString() {
-                return "PPC";
-            }
-        },
-        TASER {
-            public String toString() {
-                return "Taser";
-            }
-        },
-        FLAMER {
-            public String toString() {
-                return "Flamer";
-            }
-        },
-        MISSILE {
-            public String toString() {
-                return "Missile";
-            }
-        },
-        LRM,
-        MRM,
-        SRM,
-        PHYSICAL {
-            public String toString() {
-                return "Physical (inc. industrial equipment)";
-            }
-        },
-        AMS,
-        PRACTICAL_PHYSICAL {
-            public String toString() {
-                return "Physical (weapons only)";
-            }
-        };
-
-        public boolean matches(String name) {
-            if (name.toLowerCase().contains("ammo")) {
-                return false;
-            }
-            if (this == PHYSICAL) {
-                String lName = name.toLowerCase();
-
-                if (lName.contains("backhoe") ||
-                    lName.contains("saw") ||
-                    lName.contains("whip") ||
-                    lName.contains("claw") ||
-                    lName.contains("combine") ||
-                    lName.contains("flail") ||
-                    lName.contains("hatchet") ||
-                    lName.contains("driver") ||
-                    lName.contains("lance") ||
-                    lName.contains("mace") ||
-                    lName.contains("drill") ||
-                    lName.contains("ram") ||
-                    lName.contains("blade") ||
-                    lName.contains("cutter") ||
-                    lName.contains("shield") ||
-                    lName.contains("welder") ||
-                    lName.contains("sword") ||
-                    lName.contains("talons") ||
-                    lName.contains("wrecking")) {
-                    return true;
-                }
-            } else if (this == PRACTICAL_PHYSICAL) {
-                String lName = name.toLowerCase();
-
-                if (lName.contains("claw") ||
-                    lName.contains("flail") ||
-                    lName.contains("hatchet") ||
-                    lName.contains("lance") ||
-                    lName.contains("mace") ||
-                    lName.contains("blade") ||
-                    lName.contains("shield") ||
-                    lName.contains("sword") ||
-                    lName.contains("talons")) {
-                    return true;
-                }
-            } else if (this == MISSILE) {
-                if ((name.toLowerCase().contains("lrm") ||
-                    name.toLowerCase().contains("mrm") ||
-                    name.toLowerCase().contains("srm")) &&
-                    !name.toLowerCase().contains("ammo")) {
-                    return true;
-                }
-            } else if (this == RE_ENGINEERED) {
-                if (name.toLowerCase().contains("engineered")) {
-                    return true;
-                }
-            } else if (this == ENERGY) {
-                if (WeaponClass.LASER.matches(name) || WeaponClass.PPC.matches(name) || WeaponClass.FLAMER.matches(name)) {
-                    return true;
-                }
-            } else if (this == MACHINE_GUN) {
-                if ((name.toLowerCase().contains("mg") || name.toLowerCase().contains("machine")) && !name.toLowerCase().contains("ammo")) {
-                    return true;
-                }
-            } else if (this == BALLISTIC) {
-                return WeaponClass.AUTOCANNON.matches(name) ||
-                    WeaponClass.GAUSS.matches(name) ||
-                    WeaponClass.MISSILE.matches(name) ||
-                    WeaponClass.MACHINE_GUN.matches(name);
-            } else if (this == RAC) {
-                if (name.toLowerCase().contains("rotary")) {
-                    return true;
-                }
-            } else if (this == ULTRA) {
-                if (name.toLowerCase().contains("ultraa")) {
-                    return true;
-                }
-            } else if (name.toLowerCase().contains(this.name().toLowerCase()) && !name.toLowerCase().contains("ammo")) {
-                return true;
-            }
-            return false;
-        }
-    }
 }
